Experiments with dependency processing for changed nodes
When a new version of a node has the same location as the old version of
that node, we don't need to re-process ways and relations this node is
a member of.

See #1046

This is some preliminary experiments. It is not tested well. Currently
only works with flat node file.

Does this work in all cases? Can we be shure that the state reflected in
the middle tables/the flat node file is the same as in the output
tables? What happens if we re-start updating with an older diff file?
What happens if a nodes is deleted and then re-surrected?